+++ private/build_mutation_fields/index.js
@@ -5,7 +5,6 @@ const {
   GraphQLInputObjectType,
   GraphQLList
 } = require('graphql')
-const broadcast_resolver = require('../resolvers/broadcast.js')
 const resolver = require('../resolvers/index.js')
 const serialize_transaction_data = require('../serialize/transaction_data.js')
 const ast_to_input_types = require('./ast_to_input_types.js')
@@ -13,7 +12,6 @@ const authorization_type = require('./types/authorization_type.js')
 const configuration_default_value = require('./types/configuration_default_value.js')
 const configuration_type = require('./types/configuration_type.js')
 const packed_transaction_type = require('./types/packed_transaction_type.js')
-const transaction_receipt_type = require('./types/transaction_receipt_type.js')
 
 /**
  * Builds a GraphQL mutation input field from an `abi ast`.
@@ -22,11 +20,10 @@ const transaction_receipt_type = require('./types/transaction_receipt_type.js')
  * @name build_mutation_fields
  * @kind function
  * @param {ABI_AST} abi_ast Abstract syntax tree (AST) for a given smart contract.
- * @param {bool} broadcast Determines if transaction is pushed to the blokchain or just serialized.
  * @returns {object} GraphQL mutation fields.
  * @ignore
  */
-function build_mutation_fields(abi_ast, broadcast) {
+function build_mutation_fields(abi_ast) {
   const ast_input_object_types = ast_to_input_types(abi_ast)
 
   const fields = abi_ast.actions.reduce(
@@ -54,7 +51,7 @@ function build_mutation_fields(abi_ast, broadcast) {
 
   return {
     [abi_ast.gql_contract]: {
-      type: broadcast ? transaction_receipt_type : packed_transaction_type,
+      type: packed_transaction_type,
       description: `Update the \`${abi_ast.contract}\` smart contract.`,
       args: {
         actions: {
@@ -77,7 +74,7 @@ function build_mutation_fields(abi_ast, broadcast) {
       async resolve(
         _,
         { configuration = configuration_default_value, actions },
-        { rpc_url, private_keys = [] }
+        { rpc_url }
       ) {
         let _actions = []
         let _context_free_actions = []
@@ -118,18 +115,12 @@ function build_mutation_fields(abi_ast, broadcast) {
             configuration,
             actions: _actions,
             rpc_url,
-            private_keys,
             context_free_actions: _context_free_actions
           },
           abi_ast
         )
 
-        if (broadcast)
-          return broadcast_resolver({
-            ...packed_transaction,
-            rpc_url
-          })
-        else return packed_transaction
+        return packed_transaction
       }
     }
   }