+++ private/build_mutation_fields/transactions.js
@@ -4,35 +4,30 @@ const {
   GraphQLList,
   GraphQLNonNull
 } = require('graphql')
-const broadcast_resolver = require('../resolvers/broadcast.js')
 const resolver = require('../resolvers/index.js')
 const serialize_transaction_data = require('../serialize/transaction_data.js')
 const configuration_default_value = require('./types/configuration_default_value.js')
 const configuration_type = require('./types/configuration_type.js')
 const packed_transaction_type = require('./types/packed_transaction_type.js')
-const transaction_receipt_type = require('./types/transaction_receipt_type.js')
 
 /**
  * Generate mutation fields that can handle atomic transactions accross multiple smart contracts (1 -> x).
  * @kind function
  * @name build_transactions_mutation_fields
- * @param {object} contract_mutation_fields List of smart contract mutation fields
+ * @param {object} serialize_transaction_fields List of smart contract mutation fields
  * @param {Array<object>} abi_ast List of abstract syntax tree (AST) of the corresponding application binary interface‘s (ABI).
- * @param {bool} broadcast Determines if the transaction is pushed to the blockchain or just packed.
  * @returns {object} GraphQL mutation fields.
  * @ignore
  */
 const build_transactions_mutation_fields = (
-  contract_mutation_fields,
-  abi_ast,
-  broadcast
+  serialize_transaction_fields,
+  abi_ast
 ) =>
-  Object.keys(contract_mutation_fields).length
+  Object.keys(serialize_transaction_fields).length
     ? {
-        contract_mutations: {
-          description:
-            'Update multiple `smart contracts` in one mutation ([atomicity](https://en.wikipedia.org/wiki/Atomicity_(database_systems))).',
-          type: broadcast ? transaction_receipt_type : packed_transaction_type,
+        serialize_transaction: {
+          description: 'Serialise a list of transactions into `WASM` binary.',
+          type: packed_transaction_type,
           args: {
             actions: {
               type: new GraphQLNonNull(
@@ -40,11 +35,11 @@ const build_transactions_mutation_fields = (
                   new GraphQLNonNull(
                     new GraphQLInputObjectType({
                       name: 'transactions_type',
-                      fields: Object.keys(contract_mutation_fields).reduce(
+                      fields: Object.keys(serialize_transaction_fields).reduce(
                         (acc, key) => ({
                           ...acc,
                           [key]: {
-                            type: contract_mutation_fields[key].args.actions
+                            type: serialize_transaction_fields[key].args.actions
                               .type
                           }
                         }),
@@ -62,7 +57,7 @@ const build_transactions_mutation_fields = (
           async resolve(
             _,
             { actions, configuration = configuration_default_value },
-            { rpc_url, private_keys = [] }
+            { rpc_url }
           ) {
             let _actions = []
 
@@ -101,19 +96,12 @@ const build_transactions_mutation_fields = (
                 configuration,
                 actions: _actions,
                 rpc_url,
-                private_keys,
                 context_free_actions: _context_free_actions
               },
               abi_ast
             )
 
-            if (broadcast)
-              return broadcast_resolver({
-                ...packed_transaction,
-                private_keys,
-                rpc_url
-              })
-            else return packed_transaction
+            return packed_transaction
           }
         }
       }