+++ private/build_mutation_fields/types/packed_transaction_type.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const { GraphQLObjectType, GraphQLString, GraphQLList } = require('graphql')
-const signature_type = require('../../eosio_types/signature_type.js')
+const { GraphQLObjectType, GraphQLString } = require('graphql')
+
 /**
  * The packed transaction type.
  * @kind typedef
@@ -10,8 +10,6 @@ const signature_type = require('../../eosio_types/signature_type.js')
  * @prop {string} chain_id Hash representing the blockchain.
  * @prop {string} transaction_header Hex string representing the serialized transaction header.
  * @prop {string} transaction_body Hex string representing the serialized transaction body.
- * @prop {Array<string>} signatures List of required signatures to satisfy authorizations.
- * @prop {Array<string>} meta_signatures List of all signatures from the supplied private keys.
  */
 const packed_transaction_type = new GraphQLObjectType({
   name: 'packed_transaction',
@@ -19,7 +17,7 @@ const packed_transaction_type = new GraphQLObjectType({
   fields: () => ({
     chain_id: {
       type: GraphQLString,
-      description: 'chain id'
+      description: 'A unique ID that Specifies which chain we are operating on.'
     },
     transaction_header: {
       type: GraphQLString,
@@ -27,16 +25,7 @@ const packed_transaction_type = new GraphQLObjectType({
     },
     transaction_body: {
       type: GraphQLString,
-      description: 'Packed transaction.'
-    },
-    signatures: {
-      type: new GraphQLList(signature_type),
-      description: 'List of required signatures to satisfy authorization.'
-    },
-    meta_signatures: {
-      type: new GraphQLList(signature_type),
-      description:
-        'List of all corresponding signatures from the list of given private keys.'
+      description: 'Packed serialized transaction i.e. `wasm` instructions.'
     }
   })
 })