+++ private/eosio_types/bytes_type.js
@@ -1,15 +1,18 @@
 'use strict'
-const serialize = require('eosio-wasm-js/public/serialize/bytes.js')
+
 const { GraphQLScalarType } = require('graphql')
 
 const bytes_type = new GraphQLScalarType({
-  description: '`Bytes type`',
+  description: 'Hexedecimal text string type.',
   name: 'bytes',
   parseValue: bytes => {
     if (bytes == '') return bytes
     if (!bytes.match(/^[A-Fa-f0-9]+$/gmu))
       throw new TypeError('Invald hexadecimal string: ' + bytes)
-    return serialize(bytes)
+    if (bytes.length % 2 != 0)
+      throw new TypeError('Invalid Hexadecimal string length')
+
+    return bytes
   }
 })